refactor: Move NodeIcon into voices_node_menu
